pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "defectool"
version = "0.2.0"
description = "Point defect structure generation, relaxation and analysis"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "ase>=3.22",
    "spglib>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mace = ["mace-torch>=0.3"]
dev = ["pytest", "ruff"]

[project.scripts]
defectool = "defectool.cli:main"

[tool.setuptools.packages.find]
include = ["defectool*"]
